gpupgrade_hub: implement --daemonize
Passing the --daemonize option to gpupgrade_hub will invoke
utils.Daemonize() to split off a new background process. (The background
process knows to close its stdout/err streams because it is passed the
--daemon option.)

Any errors encountered before the hub starts will be reported with the
same output and exit code as if --daemonize were not passed, allowing
callers to wait for output.

Co-authored-by: Shoaib Lari <slari@pivotal.io>
Co-authored-by: Kevin Yeap <kyeap@pivotal.io>