XXX: First stab at synchronous operation and colored output
Use a gRPC server-side stream to communicate with the client during
cluster shutdown. For now, the status messages are tightly coupled
between the server and client, but minimal so that the client still has
control over the UI.

The shutdown operation continues whether or not the client is listening.
Errors during stream.Send() are logged but otherwise ignored.

UI-wise, output on the client side has been moved away from gplog and
towards colored status output, in the style of more modern system
managers. It's pretty hard-coded at this point.