PoC: use fsnotify to track directory changes
fsnotify is supposed to be a cross-platform Go abstraction layer around
platform-specific implementations like inotify and kqueue. This commit
creates a binary called gpwatch that can be used like so:

    gpwatch /directory/to/watch /other/directory/to/watch ...

It will monitor the passed directories, printing any detected changes,
until it is interrupted. The exit code is only zero if no changes were
detected.

Bugs:
- fsnotify itself is resource intensive on Mac, requiring an open file
  handle for every file contained in the watched directories, and this
  causes "too many open files" errors to be hit immediately when
  attempting to watch a GPDB data directory. Not super useful.
- Symlinks are not watched correctly due to the naive filepath.Walk()
  implementation. For instance, on macOS, `gpwatch /tmp` won't do
  anything.