Check that all segments are up during initialize.
Introduces new type called SegmentStatus because we don't want to
hold onto the state of "is a segment up or down" for any long period
of time. (like in a Cluster's SegConfig).

Also adds a unit test for formatting of substeps that will fail if
the substep has not been formatted yet.