ci: Publish a gpupgrade binary in the prod pipeline
A new publish-release-candidate job is added to the new prod-pipeline.yml file.
This job requires all previous jobs to have succeeded, and will then publish
an artifact to the bucket/path specified in the bin_gpupgrade_rc resource.
The existing pipeline.yml was moved to dev-pipeline.yml.
Also, the generated files now live in ci/generated.

Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>