hub: fill in s.Target only after gpinitsystem
s.Target was saved to disk during FillClusterConfigsSubStep, but it
wasn't valid for general use (only the cluster's GPHome field was
initialized).

This led to later confusion -- the version check (for GPDB less than 7)
would always succeed, because s.Target had a "version" of 0.0.0 at that
point. As another example, revert is currently unable to correctly
operate if gpinitsystem fails during initialize, because it uses
s.Target to determine whether the target cluster is up, but s.Target
isn't guaranteed to be filled in yet.

Follow the principle of "either an object is valid, or it doesn't
exist", and only initialize s.Target once gpinitsystem has been filled
in. Any parts of the target initialization code that relied on s.Target
before it was valid now take only the individual components that they
need to function. The --target-gphome config value that's passed by the
user is now saved to a top-level field in the Config struct; the
duplication is a small price to pay for sanity.